fix(wiz-docs): share trust-all TLS setup and fix hostname-verification gap
Extract applyAssistantTls() onto AIAssistantController so the health check
and the doc-search gateway configure the same trust-all SSLContext and
clear SSLParameters' endpoint identification algorithm the same way.
The gateway previously skipped that last step, so a self-signed cert
whose CN/SAN didn't match the connect host would fail there even with
chat.app.server.ssl.verify=false, while the health check succeeded.
Also closes the gateway's cached HttpClient on @PreDestroy, mirroring
AIAssistantController.closeHealthClient().
