+++ core/src/main/java/inetsoft/web/assistant/AIAssistantController.java
@@ -50,30 +50,7 @@ public class AIAssistantController {
    public void createHealthClient() {
       HttpClient.Builder builder = HttpClient.newBuilder()
          .connectTimeout(Duration.ofSeconds(3));
-
-      if(!isSslVerifyEnabled()) {
-         LOG.warn("SSL certificate verification is disabled for AI assistant health check " +
-                     "connections (chat.app.server.ssl.verify=false). Set to true in production when " +
-                     "the assistant server uses a CA-signed certificate.");
-
-         try {
-            SSLContext sslContext = SSLContext.getInstance("TLS");
-            sslContext.init(null, new TrustManager[] {
-               new X509TrustManager() {
-                  public void checkClientTrusted(X509Certificate[] chain, String authType) {}
-                  public void checkServerTrusted(X509Certificate[] chain, String authType) {}
-                  public X509Certificate[] getAcceptedIssuers() { return new X509Certificate[0]; }
-               }
-            }, null);
-            SSLParameters sslParameters = new SSLParameters();
-            sslParameters.setEndpointIdentificationAlgorithm("");
-            builder.sslContext(sslContext).sslParameters(sslParameters);
-         }
-         catch(NoSuchAlgorithmException | KeyManagementException e) {
-            LOG.warn("Could not configure trust-all SSL context for AI assistant health check", e);
-         }
-      }
-
+      applyAssistantTls(builder, "health check");
       healthClient = builder.build();
    }
 
@@ -172,6 +149,45 @@ public class AIAssistantController {
     */
    public static boolean isSslVerifyEnabled() {
       return "true".equalsIgnoreCase(SreeEnv.getProperty("chat.app.server.ssl.verify", "false"));
+   }
+
+   /**
+    * Configures an {@code HttpClient.Builder} to skip TLS certificate <em>and</em> hostname
+    * verification when {@code chat.app.server.ssl.verify} is disabled, so that private-network
+    * deployments with self-signed certificates work out of the box.
+    *
+    * <p>Public and static so every outbound connection to the assistant server — the health
+    * check here and {@code HttpAssistantDocSearchGateway}'s doc-search POST — trusts the same
+    * certificates the same way. A trust-all {@code X509TrustManager} alone only skips chain
+    * validation; without clearing {@code SSLParameters}' endpoint identification algorithm too,
+    * {@code HttpClient} still enforces hostname verification, so a self-signed certificate
+    * whose CN/SAN doesn't match the connect host would still fail here even though the operator
+    * disabled verification for exactly that case.</p>
+    *
+    * @param builder the client builder to configure, mutated in place.
+    * @param connectionLabel a short label identifying the caller, used only in the warning log
+    *                         message (e.g. {@code "health check"}, {@code "doc search"}).
+    */
+   public static void applyAssistantTls(HttpClient.Builder builder, String connectionLabel) {
+      if(isSslVerifyEnabled()) {
+         return;
+      }
+
+      LOG.warn("SSL certificate verification is disabled for AI assistant {} connections " +
+                  "(chat.app.server.ssl.verify=false). Set to true in production when the " +
+                  "assistant server uses a CA-signed certificate.", connectionLabel);
+
+      try {
+         SSLContext sslContext = SSLContext.getInstance("TLS");
+         sslContext.init(null, new TrustManager[] { TRUST_ALL }, null);
+         SSLParameters sslParameters = new SSLParameters();
+         sslParameters.setEndpointIdentificationAlgorithm("");
+         builder.sslContext(sslContext).sslParameters(sslParameters);
+      }
+      catch(NoSuchAlgorithmException | KeyManagementException e) {
+         LOG.warn("Could not configure trust-all SSL context for AI assistant " +
+                     connectionLabel, e);
+      }
    }
 
    /**
@@ -343,5 +359,21 @@ public class AIAssistantController {
    public static final String CHAT_APP_LOGO_URL = "chat.app.logo.url";
    public static final String PROXY_PATH_PREFIX = "/api/assistant/proxy";
    public static final String AI_ASSISTANT_VISIBLE = "ai.assistant.visible";
+
+   private static final X509TrustManager TRUST_ALL = new X509TrustManager() {
+      @Override
+      public void checkClientTrusted(X509Certificate[] chain, String authType) {
+      }
+
+      @Override
+      public void checkServerTrusted(X509Certificate[] chain, String authType) {
+      }
+
+      @Override
+      public X509Certificate[] getAcceptedIssuers() {
+         return new X509Certificate[0];
+      }
+   };
+
    private static final Logger LOG = LoggerFactory.getLogger(AIAssistantController.class);
 }
